Moved getIsParticipant, etc... to utils.server.ts
